Make core abstract
